4-video-editor/requirements.txt
flask==3.0.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
openai==1.12.0
requests==2.31.0
